requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.3.0
streamlit>=1.20.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
